+++ _baseStructure/features/example/delivery/handlers/exampleGet/v1_netHttp.go
@@ -2,9 +2,9 @@ package exampleGet
 
 import (
 	"encoding/json"
-	"github.com/dalikewara/ayapingping-go/v4/structure/domain"
-	"github.com/dalikewara/ayapingping-go/v4/structure/features/example/commons/presenterJSON"
-	"github.com/dalikewara/ayapingping-go/v4/structure/features/example/delivery/middlewares/checkMethod"
+	"github.com/dalikewara/ayapingping-go/v4/_baseStructure/domain"
+	"github.com/dalikewara/ayapingping-go/v4/_baseStructure/features/example/commons/presenterJSON"
+	"github.com/dalikewara/ayapingping-go/v4/_baseStructure/features/example/delivery/middlewares/checkMethod"
 	"net/http"
 )